+++ main.h
@@ -13,7 +13,7 @@ int _putstr(char *str);
 int	_put_x(unsigned int nb);
 int	_put_X(unsigned int nb);
 int get_len(unsigned int nb);
-int	_put_o(unsigned long nb);
+int	_put_o(unsigned int nb);
 int _put_u(unsigned int num);
 
 #endif
